app/requirements.txt
importlib-metadata==5.0.0
llvmlite==0.39.1
mysql-connector-python==8.0.31
numba==0.56.2
numpy==1.21.6
pandas==1.1.5
protobuf==3.20.1
py-lets-be-rational==1.0.1
py-vollib==1.0.1
py-vollib-vectorized==0.1.1
python-dateutil==2.8.2
pytz==2022.4
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
typing_extensions==4.4.0
websockets==10.3
zipp==3.9.0
